Update - 110724

+++ Frontend/src/utils/questionReqs.ts
@@ -1,18 +1,41 @@
-import { useMutation, useQuery } from "@tanstack/react-query";
+import { useInfiniteQuery, useMutation, useQuery } from "@tanstack/react-query";
 import axios from "axios";
-import {
-  AnswerType,
-  TagReturnType,
-} from "../types/global";
+import { AnswerType, TagProfileType, TagReturnType } from "../types/global";
 
-const API_URL = "https://askmeback.onrender.com/qanda";
-
-export const useTopQuestions = (userToken: string | null) => {
-  return useQuery({
-    queryKey: ["topQuestions"],
-    queryFn: async () => {
+export const useFetchInfiniteQuestions = (userToken: string) => {
+  return useInfiniteQuery({
+    queryKey: ["fetchQuestions", userToken],
+    initialPageParam: 0,
+    queryFn: async ({ pageParam }) => {
+      if (!userToken) {
+        return;
+      }
       const { data } = await axios.get(
-        "https://askmeback.onrender.com/qanda/questions/filter",
+        `${import.meta.env.VITE_APP_BACKEND_URL}/questions/getlimit/${pageParam}`,
+        {
+          headers: {
+            Authorization: `Bearer ${userToken}`,
+          },
+        }
+      );
+      return data.questions;
+    },
+    getNextPageParam: (lastPage, allPages) => {
+      return lastPage.length ? allPages.length + 1 : undefined;
+    },
+    enabled: !!userToken,
+  });
+};
+
+export const useTopQuestions = (userToken: string) => {
+  return useQuery({
+    queryKey: ["topQuestions", userToken],
+    queryFn: async () => {
+      if (!userToken) {
+        return;
+      }
+      const { data } = await axios.get(
+        import.meta.env.VITE_APP_BACKEND_URL + `/questions/filter`,
         {
           headers: {
             Authorization: `Bearer ${userToken}`,
@@ -28,10 +51,13 @@ export const useTopQuestions = (userToken: string | null) => {
 
 export const usePopularTags = (userToken: string) => {
   return useQuery({
-    queryKey: ["popularTags"],
+    queryKey: ["popularTags", userToken],
     queryFn: async () => {
+      if (!userToken) {
+        return;
+      }
       const { data } = await axios.get(
-        "https://askmeback.onrender.com/qanda/tags/popular/tags",
+        import.meta.env.VITE_APP_BACKEND_URL + `/tags/popular/tags`,
         {
           headers: {
             Authorization: `Bearer ${userToken}`,
@@ -45,15 +71,18 @@ export const usePopularTags = (userToken: string) => {
   });
 };
 
-export const questionVote = (userToken: string) => {
+export const useQuestionVote = (userToken: string) => {
   return useMutation<
     void,
     Error,
     { questionID: string; vote: number; userID: string }
   >({
     mutationFn: async ({ questionID, vote, userID }) => {
+      if (!userToken) {
+        return;
+      }
       await axios.post(
-        `${API_URL}/questions/voted`,
+        `${import.meta.env.VITE_APP_BACKEND_URL}/questions/voted`,
         {
           questionID,
           vote,
@@ -72,10 +101,13 @@ export const questionVote = (userToken: string) => {
 
 export const useSearchQuestions = (userToken: string) => {
   return useQuery({
-    queryKey: ["searchQuestions"],
+    queryKey: ["searchQuestions", userToken],
     queryFn: async () => {
+      if (!userToken) {
+        return;
+      }
       const { data } = await axios.get(
-        "https://askmeback.onrender.com/qanda/questions/searchquestions",
+        import.meta.env.VITE_APP_BACKEND_URL + `/questions/searchquestions`,
         {
           headers: {
             Authorization: `Bearer ${userToken}`,
@@ -91,12 +123,15 @@ export const useSearchQuestions = (userToken: string) => {
 
 /* ---------------------- ANSWER ------------------------------- */
 
-export const fetchAnswers = (userToken: string, questionID: string) => {
+export const useFetchAnswers = (userToken: string, questionID: string) => {
   return useQuery({
-    queryKey: ["fetchAnswers", questionID],
+    queryKey: ["fetchAnswers", questionID, userToken],
     queryFn: async () => {
+      if (!userToken || !questionID) {
+        return;
+      }
       const { data } = await axios.get<AnswerType[]>(
-        `https://askmeback.onrender.com/qanda/answers/${questionID}`,
+        import.meta.env.VITE_APP_BACKEND_URL + `/answers/${questionID}`,
         {
           headers: {
             Authorization: `Bearer ${userToken}`,
@@ -110,15 +145,18 @@ export const fetchAnswers = (userToken: string, questionID: string) => {
   });
 };
 
-export const answerVote = (userToken: string) => {
+export const useAnswerVote = (userToken: string) => {
   return useMutation<
     void,
     Error,
     { answerID: string; vote: number; userID: string }
   >({
     mutationFn: async ({ answerID, vote, userID }) => {
+      if (!userToken) {
+        return;
+      }
       await axios.post(
-        `${API_URL}/answers/voted`,
+        `${import.meta.env.VITE_APP_BACKEND_URL}/answers/voted`,
         {
           answerID,
           vote,
@@ -135,7 +173,7 @@ export const answerVote = (userToken: string) => {
   });
 };
 
-export const submitAnswer = (userToken: string) => {
+export const useSubmitAnswer = (userToken: string) => {
   return useMutation<
     void,
     Error,
@@ -151,17 +189,21 @@ export const submitAnswer = (userToken: string) => {
     }
   >({
     mutationFn: async (newAnswer) => {
-      await axios.post(`${API_URL}/answers/newanswer`, newAnswer, {
-        headers: {
-          Authorization: `Bearer ${userToken}`,
-          "X-Content-Type-Options": "nosniff",
-        },
-      });
+      await axios.post(
+        `${import.meta.env.VITE_APP_BACKEND_URL}/answers/newanswer`,
+        newAnswer,
+        {
+          headers: {
+            Authorization: `Bearer ${userToken}`,
+            "X-Content-Type-Options": "nosniff",
+          },
+        }
+      );
     },
   });
 };
 
-export const answerDelete = (
+export const useAnswerDelete = (
   answerID: string,
   userID: string,
   username: string,
@@ -169,10 +211,20 @@ export const answerDelete = (
   userToken: string
 ) => {
   return useQuery({
-    queryKey: ["answerDelete", answerID, questionID],
+    queryKey: [
+      "answerDelete",
+      answerID,
+      questionID,
+      userID,
+      userToken,
+      username,
+    ],
     queryFn: async () => {
+      if (!userToken || !userID || !username || !questionID || !answerID) {
+        return;
+      }
       const { data } = await axios.delete(
-        `https://askmeback.onrender.com/qanda/answers/deleteanswer`,
+        import.meta.env.VITE_APP_BACKEND_URL + `/answers/deleteanswer`,
         {
           data: {
             answerID: answerID,
@@ -196,12 +248,18 @@ export const answerDelete = (
 
 /* ---------------------- TAGS ------------------------------- */
 
-export const fetchTagById = (tagID: string, userToken: string) => {
+export const useFetchTagById = (
+  tagID: string | undefined,
+  userToken: string
+) => {
   return useQuery({
     queryKey: ["answerDelete", tagID, userToken],
     queryFn: async () => {
+      if (!tagID) {
+        return;
+      }
       const { data } = await axios.get<{ tag: TagReturnType[] }>(
-        `https://askmeback.onrender.com/qanda/tags/${tagID}`,
+        `${import.meta.env.VITE_APP_BACKEND_URL}/tags/${tagID}`,
         {
           headers: {
             Authorization: `Bearer ${userToken}`,
@@ -211,6 +269,28 @@ export const fetchTagById = (tagID: string, userToken: string) => {
       );
       return data.tag;
     },
-    enabled: !!userToken,
+    enabled: !!userToken && !!tagID,
+  });
+};
+
+export const useFetchUserTags = (userId: string, userToken: string) => {
+  return useQuery({
+    queryKey: ["answerDelete", userId, userToken],
+    queryFn: async () => {
+      if (!userId || !userToken) {
+        return;
+      }
+      const { data } = await axios.get<{
+        _answeredTags: TagProfileType[];
+        _questionTags: TagProfileType[];
+      }>(`${import.meta.env.VITE_APP_BACKEND_URL}/tags/user/${userId}`, {
+        headers: {
+          Authorization: `Bearer ${userToken}`,
+          "X-Content-Type-Options": "nosniff",
+        },
+      });
+      return data;
+    },
+    enabled: !!userToken && !!userId,
   });
 };
